Create The hacker quick refrence hanbook
" The Hacker’s Quick Reference handbook" authored by Hemanshu Jadhav and Bhaven 
Thalke is intended to be a comprehensive guide to cybersecurity and ethical hacking. This 
handbook aims to provide educational insights and knowledge to individuals interested in 
learning about cybersecurity concepts, practices, and techniques.
The authors strongly emphasize that the information provided in this handbook should 
only be used for ethical purposes and educational pursuits. The primary objective of this 
publication is to promote understanding and awareness of cybersecurity principles and best 
practices to enhance the security of digital systems and networks.
Readers are hereby cautioned against misusing the knowledge gained from this 
handbook for any illegal, harmful, or malicious activities. Unauthorized access to computer 
systems, networks, or any other form of unethical behaviour is strictly prohibited and may lead 
to legal consequences.
The authors, Hemanshu Jadhav and Bhaven Thalke, shall not be held liable for any 
damages, losses, or misuse of information resulting from the readers actions based on the 
content provided in this handbook.
By accessing this "Quick Reference Handbook," readers acknowledge their 
responsibility to adhere to ethical standards and use the knowledge gained responsibly and 
with integrity.
Always remember that knowledge in the field of cybersecurity comes with great 
responsibility. Let us all work together to make the digital world a safer and more secure place 
for everyone.

Bhaven Thalke
Hemanshu Jadhav